requirements.txt
python-dotenv
deepgram-sdk
aiortc
openai
